update: adjsut tests
